Small fixes to comments an types

+++ VortexEngine/src/Patterns/Multi/MeteorPattern.cpp
@@ -34,7 +34,7 @@ void MeteorPattern::blinkOff()
 
 void MeteorPattern::poststep()
 {
-  for (int meteorCount = 0; meteorCount < (LED_COUNT / 2); ++meteorCount) {
+  for (uint8_t meteorCount = 0; meteorCount < (LED_COUNT / 2); ++meteorCount) {
     // when a new meteor is created it is incerted into the stash so the blinking pattern is not interrupted
     LedPos target = (LedPos)m_randCtx.next8(LED_FIRST, LED_LAST);
     RGBColor col = m_colorset.getNext();
